remove "drop"-style compactification
This is no longer significant after the removal of xml:base. The only attribute that is left that could be "dropped" from edges to the child node when compactifying is "xsi:schemalLocation".
See the wiki issue about this: http://wiki.ruleml.org/index.php/Attribute_xsi:schemaLocation_on_Skippable_Edges