modify XSLT test script to cover maximal non-normal instance
In a maximal non-normal instance, it is possible to reverse the 
order of slotdep with slot or tupdep with tup.

Due to limitation of XSD, only RNC schemas can be used for validating
such instances. 

Therefore, the XSLT normalizer is also applied to the instances in RNC
test suites.

Not all PSOA instances have the string "PSOA" in the file name so the
condition is expanded to check for the string "Tuple".